Fix ICU to work with Android's system stlport
The version of stlport in the Android ADK includes support for RTTI, but the 
system version does not. ICU requires RTTI, so we have to explicitly depend 
upon gabi++ to provide RTTI when using the system stlport. 

Note that we can't simply include gabi++ with all modules when using the system 
stlport, as the gabi++ headers conflict with bionic when used with some other 
modules.

git-svn-id: https://src.chromium.org/chrome/trunk/deps/third_party/icu46@146346 4ff67af0-8c30-449e-8e8b-ad334ec8d88c