Use the system stdint.h on Windows.
This is now possible since VC2010 and later provide a stdint.h. VC2008
has been deprecated for building Chromium for a full year now.

There are a number of stdint.h implementations in the codebase, some of
which hide others, and most of which are not complete, and some are not
named stdint.h. The one here in icu is pwin32.h and it is the last file
which implements parts of stdint.h which prevents Blink from using the
system stdint.h.

R=jshin@chromium.org

Review URL: https://codereview.chromium.org/27079003

git-svn-id: https://src.chromium.org/chrome/trunk/deps/third_party/icu46@228516 4ff67af0-8c30-449e-8e8b-ad334ec8d88c